% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{add_extension}
\alias{add_extension}
\title{Add an extension to a filename}
\usage{
add_extension(
  filename = c("BR2_2016_S216_L001_R2_001.fastq", "foobar.fastq.gz", "fuzzbaz.txt"),
  ext = ".gz",
  no_dup = TRUE
)
}
\arguments{
\item{filename}{character, vector of one or more filenames}

\item{ext}{character, the extension to add}

\item{no_dup}{logical if TRUE then check first to determine if the filenames already
have the extension, if they do have it do not add it again}
}
\value{
the input filenames, possibly with the spcified extension added
}
\description{
Add an extension to a filename
}
