% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer.R
\name{all_orients}
\alias{all_orients}
\title{Given a primer, compute all of the possible orientations}
\usage{
all_orients(primer = "TTGAAAA-CTC-N")
}
\arguments{
\item{primer}{input sequence, see \code{\link[Biostrings]{DNAString}}}
}
\value{
name character vector
}
\description{
Given a primer, compute all of the possible orientations
}
