% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.R
\name{assign_taxonomy}
\alias{assign_taxonomy}
\title{A wrapper around assign_taxonomy that optionally populates otherwise truncated 
taxonomy levels when running straight up \code{\link[dada2]{assignTaxonomy}}.}
\usage{
assign_taxonomy(
  seqs,
  taxLevels = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
  populate_truncated = TRUE,
  truncated_value = NA_character_,
  drop_levels = c(NA_character_, "Species")[1],
  save_file = FALSE,
  filename = "taxa.csv",
  ...
)
}
\arguments{
\item{seqs}{character, see \code{\link[dada2]{assignTaxonomy}}}

\item{taxLevels}{character, \code{\link[dada2]{assignTaxonomy}}}

\item{populate_truncated}{logical, if TRUE make sure each level of \code{taxLevel} is 
represented in the returned value}

\item{truncated_value}{character or NA (default), value to assign to truncated levels.
If \code{TRUE} then \code{minBoot} is ignored - see \code{\link[dada2]{assignTaxonomy}}}

\item{drop_levels}{character, NA or a vector of levels to drop. Typical to drop 'Species'
       if \code{\link[dada2]{addSpecies}} is to be added next. If \code{TRUE} then \code{minBoot} 
       is ignored - see \code{\link[dada2]{assignTaxonomy}}
save_file logical, if TRUE save the file as CSV. If \code{TRUE} then \code{minBoot} is ignored 
       - see \code{\link[dada2]{assignTaxonomy}}
ofile character, the file to write to if save_file is TRUE}

\item{...}{further arguments for \code{\link[dada2]{assignTaxonomy}}}
}
\value{
character matrix
}
\description{
A wrapper around assign_taxonomy that optionally populates otherwise truncated 
taxonomy levels when running straight up \code{\link[dada2]{assignTaxonomy}}.
}
