% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2.R
\name{asv_fasta}
\alias{asv_fasta}
\title{Generate a FASTA object from a vector of sequences}
\usage{
asv_fasta(x, prefix = "ASV", sep = "_", pad = TRUE, file = NA)
}
\arguments{
\item{x}{character vector or a sequence count matrix with named columns}

\item{prefix}{character, by default 'ASV'}

\item{sep}{character, by default "_"}

\item{pad}{logical, if TRUE zero pad the ASV identifier count}

\item{file}{character or NA.  If not NA then save to the file specified}
}
\value{
\code{Biostrings::BStringSet}
}
\description{
Generate a FASTA object from a vector of sequences
}
