% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{audit}
\alias{audit}
\title{Provide an R session audit}
\usage{
audit(filename = "", pbs_jobid = "not known")
}
\arguments{
\item{filename}{character, the name of the file to dump to or "" to print to console}

\item{pbs_jobid}{character, the OPBS jobid if known}
}
\value{
NULL invisibly
}
\description{
Provide an R session audit
}
