% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{autopopulate_config}
\alias{autopopulate_config}
\title{Autopopulate one or more values within the config with values drawn from the config.}
\usage{
autopopulate_config(
  x = read_config("/mnt/storage/data/edna/dada/edna-dada2/config/dada2_example.yml",
    autopopulate = FALSE),
  fields = list(global = list(data_path =
    "/mnt/storage/data/edna/dada/projects/16S_Cyano_01", reference_path =
    "/mnt/storage/data/edna/dada/reference"))
)
}
\arguments{
\item{x}{configuration list}

\item{fields}{named list, provides the identifiers withion the config to be used as}
}
\description{
The second argument, \code{fields}, identifies the variables that may be found as $PLACEHOLDERS within the config file. For example, if \code{fields = list(io = c("input_directory", "output_directory"))} then we first find those values in the config file, and then search for $IO_INPUT_DIRECTORY and $IO_OUTPUT_DIRECTORY placeholders throughout the config file and replace the specified fields values.
}
