% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{check_configuration}
\alias{check_configuration}
\title{Check the configuration to make sure it is complete, borrowing from defaults as needed.}
\usage{
check_configuration(
  x = list(foo = list(bar = 7, biz = "h"), dog = list(drool = TRUE)),
  default = list(foo = list(bar = 9, biz = "p"), fish = list(pickled = "yum", breakfast
    = TRUE), dog = list(drool = TRUE, quantity = "a lot"), cat = list(hairball = TRUE))
)
}
\arguments{
\item{x}{list, configuration list}

\item{default, }{list default configuration}
}
\value{
updated configuration
}
\description{
Check the configuration to make sure it is complete, borrowing from defaults as needed.
}
