% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepairs.R
\name{count_filepairs}
\alias{count_filepairs}
\title{Count the length of filepairs}
\usage{
count_filepairs(x)
}
\arguments{
\item{x}{a list with file pairings as character vectors}
}
\value{
named numeric vector with counts for each element
}
\description{
Count the length of filepairs
}
