% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2.R
\name{count_uniques}
\alias{count_uniques}
\title{Count uniques}
\usage{
count_uniques(x, ...)
}
\arguments{
\item{x}{object from which uniques-vector can be extracted}

\item{...}{further arguments for  \code{\link[dada2]{getUniques}}}
}
\value{
integer number of uniques
}
\description{
Count uniques
}
