% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dadautils-package.R
\docType{package}
\name{dadautils-package}
\alias{dadautils-package}
\alias{dadautils}
\title{dadautils-package}
\description{
Provides added tools for working with the \href{https://benjjneb.github.io/dada2/index.html}{dada2} package in the eDNA project.
}
\seealso{
\href{https://github.com/BigelowLab/charlier}{charlier} companion package
}
