% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{default_configuration}
\alias{default_configuration}
\title{Retrieve a default configuration it's just an example}
\usage{
default_configuration()
}
\value{
configuration list
}
\description{
Retrieve a default configuration it's just an example
}
