% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepairs.R
\name{deinterleave_filepairs}
\alias{deinterleave_filepairs}
\title{Given a even numbered vector of filenames, deinterleaf into a filepairs list}
\usage{
deinterleave_filepairs(x, elements = c("forward", "reverse"))
}
\arguments{
\item{x}{a vector of items}

\item{elements}{character,the names of the elements}
}
\value{
return a filepairs list
}
\description{
Given a even numbered vector of filenames, deinterleaf into a filepairs list
}
