% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{draw_quality_profile}
\alias{draw_quality_profile}
\title{Generates a plottable object from \code{quality_profile_data} object}
\usage{
draw_quality_profile(x = quality_profile_data())
}
\arguments{
\item{x}{list as generated by \code{quality_profile_data}}
}
\value{
the input list with a "plot" element added.
}
\description{
Generates a plottable object from \code{quality_profile_data} object
}
\examples{
\dontrun{
  files <- c(system.file("extdata", "sam1F.fastq.gz", package="dada2"),
             system.file("extdata", "sam1R.fastq.gz", package="dada2"))
  result <- quality_profile_data(files)
  result_and_plot <- draw_quality_profile(result)
  print(result$plot)
 }
}
