% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{edna_path}
\alias{edna_path}
\title{Retrieve the edna storage path}
\usage{
edna_path(..., root = "/mnt/storage/projects/edna")
}
\arguments{
\item{...}{arguments for \code{\link[base]{file.path}}}

\item{root}{character, the root path}
}
\description{
Retrieve the edna storage path
}
