% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepairs.R
\name{ee_paired}
\alias{ee_paired}
\title{Compute expected errors for \code{\link[ShortRead]{alphabetScore}} vectors}
\usage{
ee_paired(x = score_pair())
}
\arguments{
\item{x}{named list of quality score vectors as returned by \code{\link[ShortRead]{alphabetScore}}}
}
\value{
named list expected errors
}
\description{
Compute expected errors for \code{\link[ShortRead]{alphabetScore}} vectors
}
\seealso{
\code{\link[dada2]{filterAndTrim}}
}
