% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepairs.R
\name{example_filepairs}
\alias{example_filepairs}
\title{Retrieve an example filepairs listing using dada2 package example data}
\usage{
example_filepairs()
}
\value{
named list of sorted forward and reverse filenames
}
\description{
Retrieve an example filepairs listing using dada2 package example data
}
