% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepairs.R
\name{expected_error_paired}
\alias{expected_error_paired}
\title{Prepare a summary of expected errors}
\usage{
expected_error_paired(filelist = example_filepairs())
}
\arguments{
\item{filelist}{named list of sorted forward and reverse filenames}
}
\value{
named list of expected error vectors (one for forward, one for reverse)
}
\description{
Prepare a summary of expected errors
}
\examples{
\dontrun{
 example_filepairs() \%>\%
   expected_error_paired()

 # $forward
 # sam1F.fastq.gz sam2F.fastq.gz 
 #              0              0 
 # 
 # $reverse
 # sam1R.fastq.gz sam2R.fastq.gz 
                0              0                
}
}
