% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{extract_sample_names}
\alias{extract_sample_names}
\title{Extract sample names fromn a list of filenames}
\usage{
extract_sample_names(
  x,
  rule = c("before first _", "before first .", "basename", "custom")[1],
  element = "forward",
  custom_fun = NULL
)
}
\arguments{
\item{x}{character, vector of one or more filenames.  This also could be a
list ala \code{\link{list_filepairs}} in which case it must have a
the specified element name \code{name}}

\item{rule}{character, the name of the rule to use}

\item{element}{character, the name of the list element used to find filenames if \code{x} is a list}

\item{custom_fun, }{a function to do custom extraction, ignored unless \code{rule} is "custom"}
}
\value{
character vector of sample names - one per input
}
\description{
Extract sample names fromn a list of filenames
}
