% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_profile.R
\name{fastq_stats}
\alias{fastq_stats}
\title{Retrieve statistics from a FastQ file}
\usage{
fastq_stats(filename = unname(unlist(example_filepairs())), n = 5e+05)
}
\arguments{
\item{filename}{character, one or more filenames}

\item{n}{(Optional). Default 500,000.
The number of records to sample from the fastq file.}
}
\value{
tibble 
  \describe{
     \item{filename}{character, the name of the file}
     \item{nread numeric}{the number of reads in the file}
     \item{ncycle}{list column of numeric vectors, the number of cycles per read}
     \item{quantile_cycles}{list columns of numeric vector, quantiles of ncycles from 0.01 to 0.99 by 0.01}
     \item{per_cycle_quality}{list columns of tibbles, quality per cycle metrics}
   }
}
\description{
Retrieve statistics from a FastQ file
}
