% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figaro.R
\name{figaro_defaults}
\alias{figaro_defaults}
\title{Retrieve default values for figaro arguments}
\usage{
figaro_defaults()
}
\value{
table of 
\itemize{
 \item{name the verbose name of the argument}
 \item{shortname the shortname of the argumement}
 \item{value the default value}
}
}
\description{
Retrieve default values for figaro arguments
}
