% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2.R
\name{filter_and_trim}
\alias{filter_and_trim}
\title{Filter and trim}
\usage{
filter_and_trim(
  filelist,
  output_path = file.path(dirname(filelist$forward[1]), "filterAndTrim"),
  save_results = FALSE,
  ...
)
}
\arguments{
\item{filelist}{list of forward and reverse fastq files}

\item{output_path}{character, the output path}

\item{save_results}{logical, save CSV if TRUE}

\item{...}{other arguments for \code{\link[dada2]{filterAndTrim}}}
}
\value{
integer matrix as tibble see \code{\link[dada2]{filterAndTrim}}
}
\description{
Filter and trim
}
