% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_configuration}
\alias{get_configuration}
\title{Get the configuration use default if needed}
\usage{
get_configuration(x, default = default_configuration(), check = FALSE)
}
\arguments{
\item{x}{configuration filename}

\item{default}{list describing the default}

\item{check}{logical, if TRUE check the config against the default}
}
\value{
list config values or a try-error
}
\description{
Get the configuration use default if needed
}
