% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_pbs_jobid}
\alias{get_pbs_jobid}
\title{Retrieve the PBS JIOBID if available}
\usage{
get_pbs_jobid(missing_text = "not known")
}
\arguments{
\item{missing_text}{character, the text to return if now jobid found'}
}
\value{
character jobid or missing_text value
}
\description{
Retrieve the PBS JIOBID if available
}
