% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{identify_upgrades}
\alias{identify_upgrades}
\title{Retrieve a table of packages suitable for updating}
\usage{
identify_upgrades(lib_pattern = "^/mnt/modules/bin/dada2")
}
\arguments{
\item{lib_pattern}{character, A glob pattern specification that is used to filter
library locations}
}
\value{
a table of packages identified for update
}
\description{
Retrieve a table of packages suitable for updating
}
