% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{installed_packages}
\alias{installed_packages}
\title{A wrapper around \code{installed.packages} the returns a table}
\usage{
installed_packages(...)
}
\arguments{
\item{...}{more arguments for \code{installed.packages}}
}
\value{
table
}
\description{
A wrapper around \code{installed.packages} the returns a table
}
