% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepairs.R
\name{interleave_filepairs}
\alias{interleave_filepairs}
\title{Given a filepair list, make a vector of alternating filenames}
\usage{
interleave_filepairs(x)
}
\arguments{
\item{x}{a filepairs list (two element list for forard and reverse filenames)}
}
\value{
see \code{\link[auntie]{interleave_filepairs}}
}
\description{
Given a filepair list, make a vector of alternating filenames
}
