% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2.R
\name{learn_errors}
\alias{learn_errors}
\title{Run dada2::learnErrors on a set of fastq files}
\usage{
learn_errors(
  filelist,
  ...,
  output_path = dirname(filelist$forward[1]),
  save_output = FALSE,
  save_graphics = FALSE
)
}
\arguments{
\item{filelist}{list of forward and reverse fastq files}

\item{...}{arguments for \code{\link[dada2]{learnErrors}}}

\item{output_path}{character, the output path}

\item{save_output}{logical, if TRUE save the output to the specified output_path}

\item{save_graphics}{logical, if TRUE try to capture quality plots form the resulting cut_files}
}
\description{
Run dada2::learnErrors on a set of fastq files
}
