% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{list_fastq}
\alias{list_fastq}
\title{List fastq files and separate into forward and reverse reads}
\usage{
list_fastq(
  path,
  pattern_forward = "^.*_R1_001\\\\.fastq",
  pattern_reverse = "^.*_R2_001\\\\.fastq"
)
}
\arguments{
\item{path}{character, the input path}

\item{pattern_forward}{file pattern}

\item{pattern_reverse}{file pattern}
}
\value{
named list of sorted foreward and reverse fastq filenames
}
\description{
List fastq files and separate into forward and reverse reads
}
