% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{logger_levels}
\alias{logger_levels}
\title{Retrieve an integer vector of known logger levels for \code{futile.logger}}
\usage{
logger_levels()
}
\value{
a named integer vector
}
\description{
Retrieve an integer vector of known logger levels for \code{futile.logger}
}
