% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{make_path}
\alias{make_path}
\title{Given a path - make it if it doesn't exist}
\usage{
make_path(path)
}
\arguments{
\item{path}{character, the path to check and/or create}
}
\value{
logical, TRUE if the path exists
}
\description{
We don't use R's \code{dir.create} in case we a reaching across volumes which 
has caused headaches at time.
}
