% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.R
\name{merge_taxonomy}
\alias{merge_taxonomy}
\title{Combine fasta sequence, taxonomy and bootstrap values}
\usage{
merge_taxonomy(fasta, taxa, filename = c(NA, "ASV_taxa.csv")[1])
}
\arguments{
\item{fasta}{FASTA sequence data}

\item{taxa}{list, as produced by \code{\link{assign_taxonomy}}}

\item{filename}{character, filename to write the table to, or NA to skip writing}
}
\value{
tibble of ASV taxonomy
}
\description{
Combine fasta sequence, taxonomy and bootstrap values
}
