% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{old_packages}
\alias{old_packages}
\title{A wrapper around \code{old.packages} the returns a table}
\usage{
old_packages(...)
}
\arguments{
\item{...}{more arguments for \code{old.packages}}
}
\value{
table
}
\description{
A wrapper around \code{old.packages} the returns a table
}
