% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepairs.R
\name{paired_ee_per_read}
\alias{paired_ee_per_read}
\title{Compute EE per read for \code{\link[Rsubread]{qualityScores}} (one or both may be empty)}
\usage{
paired_ee_per_read(x = paired_quality_scores(), fun = sum, na.rm = TRUE)
}
\arguments{
\item{x}{paired list of quality scores as per \code{quality_scores}}

\item{fun}{the function to use for summary (by default sum)}

\item{na.rm}{logical, if TRUE then remove NAs before computing \code{fun}}
}
\value{
paired list of numeric vectors with quality per read or NULL where inputs are empty
}
\description{
Compute EE per read for \code{\link[Rsubread]{qualityScores}} (one or both may be empty)
}
