% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepairs.R
\name{paired_quality_scores}
\alias{paired_quality_scores}
\title{Compute quality scores using \code{\link[Rsubread]{qualityScores}}}
\usage{
paired_quality_scores(filelist = example_filepairs(), nreads = -1, ...)
}
\arguments{
\item{filelist}{a paired list forward/reverse filenames}

\item{nreads}{numeric, see \code{\link[Rsubread]{qualityScores}}}

\item{...}{other arguments for \code{\link[Rsubread]{qualityScores}}}
}
\value{
a paired list of quality score matrices
}
\description{
Compute quality scores using \code{\link[Rsubread]{qualityScores}}
}
