% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2.R
\name{plot_qualityProfiles}
\alias{plot_qualityProfiles}
\title{Plot a suite of quality profiles. Similar to \code{\link{plot_qualityProfile}}
but works with paired forward/reverse files.}
\usage{
plot_qualityProfiles(x, n = 2, filename = "quality_profiles.pdf")
}
\arguments{
\item{x}{a list with file pairings as character vectors}

\item{n}{integer, the number of profiles (forward and reverse) to plot}

\item{filename}{character, the name of the output PDF file or NULL to skip file handling}
}
\description{
Plot a suite of quality profiles. Similar to \code{\link{plot_qualityProfile}}
but works with paired forward/reverse files.
}
