% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer.R
\name{primer_hits}
\alias{primer_hits}
\title{Count the number of matches of the primer in the subject}
\usage{
primer_hits(primer, fn)
}
\arguments{
\item{primer}{character, pattern string}

\item{fn}{XString subject to match}
}
\value{
numeric, count of hits
}
\description{
Count the number of matches of the primer in the subject
}
