% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{purge_files}
\alias{purge_files}
\title{Purge files froma folder bearing a name pattern}
\usage{
purge_files(
  path = ".",
  pattern = "^\\\\.Rsubread_qualityScores_score_pid",
  glob = FALSE
)
}
\arguments{
\item{path}{character, the path to search}

\item{pattern}{character, pattern as a regex or if \code{glob = TRUE} then a glob pattern}

\item{glob}{logical, if TRUE the pattern is a glob like "foo.*" and not regular expression}
}
\value{
named logical vector where TRUE indicates a file was removed, possibly empty
}
\description{
Purge files froma folder bearing a name pattern
}
