% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_profile.R
\name{quality_profile}
\alias{quality_profile}
\title{Compute quality profiles for one or more FastQ files.}
\usage{
quality_profile(
  fl = c(system.file("extdata", "sam1F.fastq.gz", package = "dada2"),
    system.file("extdata", "sam1R.fastq.gz", package = "dada2")),
  n = 5e+05,
  aggregate = FALSE
)
}
\arguments{
\item{fl}{(Required). \code{character}.
File path(s) to fastq or fastq.gz file(s).}

\item{n}{(Optional). Default 500,000.
The number of records to sample from the fastq file.}

\item{aggregate}{(Optional). Default FALSE.
If TRUE, compute an aggregate quality profile for all fastq files provided.}
}
\value{
as list with the following elements
\itemize{
  \item{files character vector of input filenames}
  \item{opts list of inut arguments}
  \item{stat_data} table of statistical results
  \item{ann_data} table of annotation info
  \item{plot_data} table of stats transformed for plotting
  \item{plot_summary} table of summary stats for plotting
  \item{plot} ggplot2 object suitable for printing
}
}
\description{
This a thin wrapper around \code{\link[dada2]{plotQualityProfile}} but produces a list
that includes tables of ancillary statistics in addition to a printable plot object.
}
\examples{
\dontrun{
  result <- quality_profile(system.file("extdata", "sam1F.fastq.gz", package="dada2"))
  print(result$plot)
 }
}
\seealso{
\code{\link[dada2]{plotQualityProfile}}
}
