% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_profile.R
\name{quality_profile_drawing}
\alias{quality_profile_drawing}
\title{Generates a plottable object from \code{quality_profile_data} object}
\usage{
quality_profile_drawing(x = quality_profile_data())
}
\arguments{
\item{x}{list as generated by \code{quality_profile_data}}
}
\value{
a ggplot object
}
\description{
This is not pair savvy.  Plots are faceted by files in the order the come in \code{x$files}
}
\examples{
\dontrun{
  files <- c(system.file("extdata", "sam1F.fastq.gz", package="dada2"),
             system.file("extdata", "sam1R.fastq.gz", package="dada2"))
  result <- quality_profile_data(files)
  result <- quality_profile_drawing(result)
  print(result)
 }
}
