% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_profile.R
\name{quality_profile_pairs_drawing}
\alias{quality_profile_pairs_drawing}
\title{Generates a plottable objects from \code{quality_profile_data} object by filepair (forward/reverse)}
\usage{
quality_profile_pairs_drawing(X)
}
\arguments{
\item{X}{named two element list of profile - "forward" and "reverse"}
}
\value{
list ggplot objects
}
\description{
This is pair savvy.  Plots are faceted by files in the order the come in each element of \code{x$files}
(In theory they will be paired horizontally [forward, reverse])
}
