% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{read_config}
\alias{read_config}
\title{Read the configuration}
\usage{
read_config(filename, autopopulate = TRUE, autodetect_cores = TRUE)
}
\arguments{
\item{filename}{configuration filename}

\item{autopopulate}{logical, if TRUE then autopopulate the configuration}

\item{autodetect_cores}{logical, if TRUE and multithread is present (and "auto")
then autoset the number of cores.}
}
\value{
list config values or a try-error
}
\description{
Read the configuration
}
