% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{read_errors}
\alias{read_errors}
\title{Read errors object}
\usage{
read_errors(filename = "learn_errors.rds")
}
\arguments{
\item{filename}{the name of the file to read}
}
\value{
errors object
}
\description{
Read errors object
}
