% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepairs.R
\name{read_fastq_paired}
\alias{read_fastq_paired}
\title{Read fastq files in a filepairs listing}
\usage{
read_fastq_paired(filelist = example_filepairs())
}
\arguments{
\item{filelist}{named list of sorted forward and reverse filenames (reverse may be empty)}
}
\value{
named list of \code{\link[ShortRead]{ShortReadQ}} objects for foreward and reverse
}
\description{
Read fastq files in a filepairs listing
}
