% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_mergers}
\alias{read_mergers}
\title{Read mergers object}
\usage{
read_mergers(filename = "mergers")
}
\arguments{
\item{filename}{the name of the file to read}
}
\value{
\code{\link{merge_pairs}} listing
}
\description{
Read mergers object
}
