% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_QPP}
\alias{read_QPP}
\title{Read quality_profile_pairs object}
\usage{
read_QPP(filename = "qpp")
}
\arguments{
\item{filename}{the name of the file to read}
}
\value{
\code{\link{quality_profile_pairs}} listing
}
\description{
Read quality_profile_pairs object
}
