% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_RDS}
\alias{read_RDS}
\title{A wrapper around \code{\link[base]{readRDS}}}
\usage{
read_RDS(file, ...)
}
\arguments{
\item{file}{the filename to read}

\item{...}{other arguments for \code{\link[base]{readRDS}}}
}
\value{
R object
}
\description{
A wrapper around \code{\link[base]{readRDS}}
}
