% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_profile.R
\name{reads_overlap}
\alias{reads_overlap}
\title{Compute expected overlap of read pairs}
\usage{
reads_overlap(f = 250, r = 189, a = 400)
}
\arguments{
\item{f}{numeric, length of forward read}

\item{r}{numeric, length of reverse read}

\item{a}{numeric, expected amplicon length}
}
\value{
numeric, estimated overlap of reads
}
\description{
Compute expected overlap of read pairs
}
