% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer.R
\name{remove_primers}
\alias{remove_primers}
\title{Remove primers}
\usage{
remove_primers(
  filelist,
  output_path = file.path(dirname(filelist$forward[1]), "no_primers"),
  save_results = TRUE,
  compress = TRUE,
  ...
)
}
\arguments{
\item{filelist}{list of forward and reverse fastq files}

\item{output_path}{character, the output path}

\item{save_results}{logical, save CSV if TRUE}

\item{compress}{logical, see \code{\link[dada2]{removePrimers}}}

\item{...}{other arguments for \code{\link[dada2]{removePrimers}}}
}
\value{
two element list of tibbles see \code{\link[dada2]{removePrimers}}
}
\description{
Remove primers
}
