% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.R
\name{run_cutadapt}
\alias{run_cutadapt}
\title{Run cutadapt}
\usage{
run_cutadapt(
  cut_files,
  filt_files,
  CFG,
  save_output = TRUE,
  save_graphics = FALSE
)
}
\arguments{
\item{cut_files}{two element list for forward and reverse cutadapt results to be created}

\item{filt_files}{two element list of forward and reverse filtered files}

\item{CFG}{list of configuration}

\item{save_output}{logical, if TRUE try to capture the output of cutadapt to text files,
one per cutadapt file generated, in the cutadapt output directory}

\item{save_graphics}{logical, if TRUE try to capture quality plots form the resulting cut_files}
}
\value{
numeric codes, one per cut_file pairing as per \code{system2} where 0 means success
}
\description{
Run cutadapt
}
