% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2.R
\name{run_dada}
\alias{run_dada}
\title{Run dada}
\usage{
run_dada(filelist, errs, ...)
}
\arguments{
\item{filelist}{list of forward and reverse fastq files}

\item{errs}{list of forward and reverse outputs of learnErrors}

\item{...}{arguments for \code{\link[dada2]{dada}}}
}
\value{
list with elements for forward and reverse as returned by \code{\link[dada2]{dada}}
}
\description{
Run dada
}
