% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figaro.R
\name{run_figaro}
\alias{run_figaro}
\title{Run figaro}
\usage{
run_figaro(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{table of arguments - see \code{\link{figaro_defaults}}}

\item{verbose}{logical, if TRUE output the generated command string}

\item{...}{other arguments for \code{\link[base]{system2}}}
}
\value{
hmmm
}
\description{
Run figaro
}
