% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2.R
\name{seqtab_to_matrix}
\alias{seqtab_to_matrix}
\title{Convert an eDNA sequence counts table to a \code{dada2} friendly matrix}
\usage{
seqtab_to_matrix(x)
}
\arguments{
\item{x}{filename or table, if a filename we try to read it as CSV}
}
\value{
\code{dada2} friendly matrix
}
\description{
Convert an eDNA sequence counts table to a \code{dada2} friendly matrix
}
