% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2.R
\name{sequence_lengths}
\alias{sequence_lengths}
\title{Count the charcater lengths of sequences}
\usage{
sequence_lengths(x, ofile = c(NA, "sequence_lengths.csv")[1])
}
\arguments{
\item{x}{matrix table of sequences as per \code{\link[dada2]{removeBimeraDenovo}}}

\item{ofile}{character or NA, if not NA save result to a CSV file}
}
\value{
character counts
}
\description{
Count the charcater lengths of sequences
}
