% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{strip_extension}
\alias{strip_extension}
\title{Strip the extension(s) off of a filename}
\usage{
strip_extension(
  filename = c("BR2_2016_S216_L001_R2_001.fastq", "foobar.fastq.gz", "fuzzbaz.txt"),
  ext = ".fastq"
)
}
\arguments{
\item{filename}{character one or more filenames}

\item{ext}{character, one or more extension patterns}
}
\value{
filename with extension stripped
}
\description{
Note if ext is ".fastq" then ".fastq.gz" and ".fastq.tar.gz" will also be stripped
}
