% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2.R
\name{table_as_matrix}
\alias{table_as_matrix}
\title{Convert a table (like a tibble or data.frame) to a matrix}
\usage{
table_as_matrix(x, rowname = 1, transpose = TRUE)
}
\arguments{
\item{x}{table object like a tibble or data.frame}

\item{rowname}{the column (variable) to use as source of rownames (before optional transpose)}

\item{transpose}{logical, if TRUE transpose just before returning.}

\item{a}{matrix with column and row names}
}
\description{
Useful for reconstructing a \code{dada2} matrix from an eDNA table
}
