% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.R
\name{taxa_remove}
\alias{taxa_remove}
\title{Filter by removal of one or more instances of values found in one or more variables.}
\usage{
taxa_remove(
  x,
  vars = list(any = "Chloroplast"),
  taxa_levels = c("Kingdom", "Supergroup", "Division", "Class", "Order", "Family",
    "Genus", "Species"),
  index = TRUE
)
}
\arguments{
\item{x}{a table as produced by \code{\link[dada2]{assignTaxonomy}} or \code{\link[dada2]{addSpecies}}}

\item{vars}{named list or NULL, with one character vector per variable (column) specifying 
the instances to remove.  Names of the list must be present in the variable (column) 
names of the input ar \code{any} in which case the values are searched for in any 
variable.  It is not an error to request filtering of a non-existant column, but 
a warning is generated. If NULL then return the input as a table.}

\item{taxa_levels}{character, a vector of the column names to search}

\item{index}{logical, if TRUE return a logical index into the original input where TRUE
indicates rows to be removed.  If FALSE then return a filtered table.}
}
\value{
a table with records (rows) removed or a logical index vector
}
\description{
Filter by removal of one or more instances of values found in one or more variables.
}
