% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2.R
\name{taxtable_to_matrix}
\alias{taxtable_to_matrix}
\title{Convert an eDNA taxonomy table to a \code{dada2} friendly matrix}
\usage{
taxtable_to_matrix(x, ...)
}
\arguments{
\item{x}{filename or table, if a filename we try to read it as CSV}

\item{...}{further arguments for \code{table_to_matrix}}
}
\value{
\code{dada2} friendly matrix
}
\description{
Convert an eDNA taxonomy table to a \code{dada2} friendly matrix
}
