% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{verify_filepairs}
\alias{verify_filepairs}
\title{Verify that a list of file pairs is suitable for processing}
\usage{
verify_filepairs(x, elements = c("forward", "reverse"))
}
\arguments{
\item{x}{a list with file pairings as character vectors}

\item{elements}{character, the names of the file pair elements to test}
}
\value{
the input list
}
\description{
Verify that a list of file pairs is suitable for processing
}
