% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.R
\name{verify_taxalevels}
\alias{verify_taxalevels}
\title{Verify taxa levels match depoth of taxa in reference database.}
\usage{
verify_taxalevels(taxaLevels, refFasta, delim = ";")
}
\arguments{
\item{taxaLevels}{character, vector fo desired taxa levels}

\item{refFasta}{character, filename of the refernce database}

\item{delim}{character, the delimiter used to split the first contig description line in the refFasta}
}
\value{
logical TRUE if taxaLevels and refdb have the same depth of classifications
}
\description{
Verification is by comparing count of taxa levels with the first line in reference
database which, in dada2-world, is a delimited list of levels.
}
