% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{write_errors}
\alias{write_errors}
\title{Write errors object}
\usage{
write_errors(x, filename = "learn_errors.rds")
}
\arguments{
\item{x}{errors object from \code{\link{learn_errors}}}

\item{filename}{the name of the file to write}
}
\value{
the input errors object
}
\description{
Write errors object
}
