% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_mergers}
\alias{write_mergers}
\title{Write mergers object}
\usage{
write_mergers(x, filename = "mergers")
}
\arguments{
\item{x}{output from \code{\link{merge_pairs}}}

\item{filename}{the name of the file to write}
}
\value{
the input errors object
}
\description{
Write mergers object
}
