% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_RDS}
\alias{write_RDS}
\title{A tidy wrapper around \code{\link[base]{saveRDS}}}
\usage{
write_RDS(object, file = "", ...)
}
\arguments{
\item{object}{the R object to be saved}

\item{file}{character or file connection}

\item{...}{other arguments for \code{\link[base]{saveRDS}}}
}
\value{
the original object invisibly
}
\description{
A tidy wrapper around \code{\link[base]{saveRDS}}
}
\examples{
\dontrun{
  x <- my_big_process() \%>\%
    write_RDS("my_big_data.rds") \%>\%
    dplyr::filter(foo <= 7) \%>\%
    readr::write_csv("my_smaller_data.csv")
}
}
