% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dadautils-package.R
\name{dadautils-deprecated}
\alias{dadautils-deprecated}
\title{Deprecated functions in dadautils}
\description{
These functions still work but should be avoided.
}
\details{
\itemize{
 \item \code{\link{list_fastq}}: Please use \code{\link{list_filepairs}}
}
}
