% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dadautils-package.R
\docType{package}
\name{dadautils-package}
\alias{dadautils-package}
\alias{dadautils}
\title{dadautils-package}
\description{
Provides added tools for working with the [dada2](https://benjjneb.github.io/dada2/index.html) package in the eDNA project.
}
