% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepairs.R
\name{deinterleave_filepairs}
\alias{deinterleave_filepairs}
\title{Given a even numbered vector of filenames, deinterleaf into a filepairs list}
\usage{
deinterleave_filepairs(x, ...)
}
\arguments{
\item{x}{a vector of items}

\item{...}{arguments for \code{\link[auntie]{deinterleave_filepairs}}}
}
\value{
see \code{\link[auntie]{deinterleave_filepairs}}
}
\description{
Given a even numbered vector of filenames, deinterleaf into a filepairs list
}
