% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepairs.R
\name{expected_error}
\alias{expected_error}
\title{Compute expected errors from a vector - presumably \code{\link[ShortRead]{alphabetScore}} vector}
\usage{
expected_error(x)
}
\arguments{
\item{x}{numeric vector}

\item{method}{charcater by default 'filtAndTrim' from \code{\link[dada2]{filterAndTrim}} description of 
\code{maxEE}}
}
\value{
numeric expected error
}
\description{
Compute expected errors from a vector - presumably \code{\link[ShortRead]{alphabetScore}} vector
}
\seealso{
\code{\link[dada2]{filterAndTrim}}
}
