% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{extract_sample_names}
\alias{extract_sample_names}
\title{Extract sample names fromn a list of filenames}
\usage{
extract_sample_names(x, rule = "before first _")
}
\arguments{
\item{x}{character, vector of one or more filenames.  This also could be a
list ala \code{\link{list_filepairs}} in which case it must have a
the specified element name \code{name}}

\item{rule}{character, the name of the rule to use}

\item{name}{character, the name of the liust element to find filenames if \code{x} is a list}
}
\value{
character vector of sample names - one per input
}
\description{
Extract sample names fromn a list of filenames
}
