% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2.R
\name{filter_and_trim}
\alias{filter_and_trim}
\title{Filter and trim}
\usage{
filter_and_trim(
  filelist,
  output_path = file.path(dirname(filelist$forward[1]), "filterAndTrim"),
  save_results = FALSE,
  compress = TRUE,
  multithread = charlier::count_cores(),
  truncLen = "auto",
  cutoff_params = list(score = 30, model = "Mean ~ poly(Cycle, 2)"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{filelist}{list of forward and reverse fastq files}

\item{output_path}{character, the output path}

\item{save_results}{logical, save CSV if TRUE}

\item{compress}{logical, see \code{\link[dada2]{filterAndTrim}}}

\item{multithread}{numeric, the number of cores to use. Defaults to \code{\link{count_cores}}}

\item{truncLen}{character or numeric.  If character and "auto" then auto-compute trunLen, default "auto"}

\item{cutoff_params}{list, in the even that trunLen is "auto" then pass these params to 
\code{\link{quality_profile_cutoff}}}

\item{verbose}{logical, if TRUE output informational messages}

\item{...}{other arguments for \code{\link[dada2]{filterAndTrim}}}
}
\value{
integer matrix as tibble see \code{\link[dada2]{filterAndTrim}}
}
\description{
Filter and trim
}
