% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{list_filepairs}
\alias{list_filepairs}
\title{List files and separate into forward and reverse file lists based upon filenaming patterns}
\usage{
list_filepairs(
  path,
  pattern_forward = "*_R1_001.fastq",
  pattern_reverse = "*_R2_001.fastq",
  glob = TRUE,
  ...
)
}
\arguments{
\item{path}{character, the input path}

\item{pattern_forward}{file pattern}

\item{pattern_reverse}{file pattern}

\item{glob}{logical, if \code{TRUE} the input patterns are considered file globs like "*_R1_001.fastq" and will be converted to regex patterns using \code{\link[utils]{glob2rx}}.  If glob is \code{FALSE} then the the patterns are passed to \code{\link[base]{list.files}} as provided by the user.}

\item{...}{further arguments for \code{\link[base]{list.files}}}
}
\value{
named list of sorted foreward and reverse filenames
}
\description{
List files and separate into forward and reverse file lists based upon filenaming patterns
}
