% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{list_filepairs}
\alias{list_filepairs}
\title{List files and separate into forward and reverse file lists based upon filenaming patterns.}
\usage{
list_filepairs(
  path,
  pattern_forward = "^.*R1_001",
  pattern_reverse = "^.*R2_001",
  glob = FALSE,
  verify = TRUE,
  ...
)
}
\arguments{
\item{path}{character, the input path}

\item{pattern_forward}{file pattern, the default is to match either "_R1_001.fastq" or "_R1_001.fastq.gz"}

\item{pattern_reverse}{file pattern, the default is to match either "_R2_001.fastq" or "_R2_001.fastq.gz"}

\item{glob}{logical, if \code{TRUE} the input patterns are considered file globs like "*_R1_001.fastq"
and will be converted to regex patterns using \code{\link[utils]{glob2rx}}.
If glob is \code{FALSE} then the the patterns are passed to \code{\link[base]{list.files}}
as provided by the user.}

\item{verify}{logical, if TRUE test that the filepairs are the same length}

\item{...}{further arguments for \code{\link[base]{list.files}}}
}
\value{
named list of sorted foreward and reverse filenames (reverse possibly zero-length character)
}
\description{
If reverse files are not present (ala \code{PacBio}) then that element is set to a zero-length
character element.
}
