% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepairs.R
\name{list_filepairs}
\alias{list_filepairs}
\title{List files and separate into forward and reverse file lists based upon filenaming patterns.}
\usage{
list_filepairs(path, ...)
}
\arguments{
\item{path}{character, the input path}

\item{...}{arguments for \code{\link[auntie]{list_filepairs}}}
}
\value{
see \code{\link[auntie]{list_filepairs}}
}
\description{
If reverse files are not present (ala \code{PacBio}) then that element is set to a zero-length
character element.
}
