% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2.R
\name{merge_pairs}
\alias{merge_pairs}
\title{Merge pairs ala dada2::mergePairs}
\usage{
merge_pairs(filelist, dada_r, ...)
}
\arguments{
\item{filelist}{list of forward and reverse fastq files}

\item{dada_r}{list of dada2::dada results}

\item{...}{arguments for \code{\link[dada2]{mergePairs}}}
}
\value{
as returned by \code{\link[dada2]{mergePairs}}
}
\description{
Merge pairs ala dada2::mergePairs
}
