% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepairs.R
\name{paired_ee_threshold}
\alias{paired_ee_threshold}
\title{Threshold paired expected error computations}
\usage{
paired_ee_threshold(
  x = paired_ee_per_read(),
  thresholds = c(1, 2, 3, 4, 5),
  op = `<=`
)
}
\arguments{
\item{x}{paried list of expected errors - see \code{paired_ee_per_read}}

\item{thresholds}{numeric vector of one or more thresholds}

\item{op}{the comparative operator to use for the thrrsholding process (backquoted)}
}
\value{
a two element list of tibbles with threshold counts of errors-per-read above the threshold
}
\description{
Threshold paired expected error computations
}
