% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2.R
\name{plot_qualityProfile}
\alias{plot_qualityProfile}
\title{Plot quality profiles for one or more FASTQs}
\usage{
plot_qualityProfile(x, filename = c(NA, "qualityProfile.pdf")[1], ...)
}
\arguments{
\item{x}{character vector of fastq filenames}

\item{filename}{character, the name of the PDF file to generate or NA}

\item{...}{further arguments for \code{\link[dada2]{plotQualityProfile}}}
}
\description{
Plot quality profiles for one or more FASTQs
}
