% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_profile.R
\name{quality_profile_cutoff}
\alias{quality_profile_cutoff}
\title{Compute cutoff levels \code{trunLen} for \code{\link[dada2]{filterAndTrim}}}
\usage{
quality_profile_cutoff(
  x = quality_profile_data(),
  method = "ruler",
  params = list(score = 30, model = "Mean ~ poly(Cycle, 2)", quantile_min = 0.99),
  form = c("full", "reduced")[2]
)
}
\arguments{
\item{x}{list of quality score data as per \code{quality_profile_data}}

\item{method}{character the threshold method to use - just "ruler" method now}

\item{params}{list (or NULL) of parameters needed to implement the specified method 
\itemize{
  \item{score numeric,  the threshold score to use with "ruler" method}
  \item{model character, the model}
  \item{quantile_min numeric, quantile minimum used as a back stop if ruler method fails, set to NA to skip.  Only
        used if form is "reduced".}
}}

\item{form}{character either "full" or "reduced".  If reduce return just the threshold row.}
}
\value{
the input list with an added cutoff tibble of 
\itemize{
 \item{Cycle the predicted Cycle value - the computed cutoff}
 \item{Score the fitted Score value}
 \item{status "a" for autothreshold, "f" for fixed threshold, "p" for percentile threshold}
 \item{model the model as a list column (for example an lm model)}
 \item{file the name of the file}
 }
}
\description{
Compute cutoff levels \code{trunLen} for \code{\link[dada2]{filterAndTrim}}
}
\examples{
\dontrun{
  files <- c(system.file("extdata", "sam1F.fastq.gz", package="dada2"),
             system.file("extdata", "sam1R.fastq.gz", package="dada2"))
  result <- quality_profile_data(files) \%>\%
   quality_profile_cutoff()
 }
}
