% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepairs.R
\name{score_paired}
\alias{score_paired}
\title{Compute scores for \code{\link[ShortRead]{ShortReadQ}} objects listing}
\usage{
score_paired(x = read_fastq_paired())
}
\arguments{
\item{x}{named list of sorted forward and reverse \code{\link[ShortRead]{ShortReadQ}} objects}
}
\value{
named list of quality score vectors as returned by \code{\link[ShortRead]{alphabetScore}}
}
\description{
Compute scores for \code{\link[ShortRead]{ShortReadQ}} objects listing
}
\seealso{
\code{\link[ShortRead]{alphabetScore}}
}
