% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepairs.R
\name{size_filepairs}
\alias{size_filepairs}
\title{Compute and test the size of filepairs}
\usage{
size_filepairs(x, ...)
}
\arguments{
\item{x}{named list of filepairs, typically 'forward' and 'reverse'}

\item{...}{arguments for \code{\link[auntie]{size_filepairs}}}
}
\value{
see \code{\link[auntie]{size_filepairs}}
}
\description{
Compute and test the size of filepairs
}
