% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.R
\name{taxa_remove}
\alias{taxa_remove}
\title{Filter by removal of one or more instances of values found in one or more variables.}
\usage{
taxa_remove(x, vars = list(Order = "Chloroplast"))
}
\arguments{
\item{x}{a table as produced by \code{\link[dada2]{assignTaxonomy}} or \code{\link[dada2]{addSpecies}}}

\item{vars}{named list or NULL, with one character vector per variable (column) specifying the instances to remove.  Names of the list must be present in the variable (column) names of the input.  If NULL then return the input as a table.}
}
\value{
a table with records (rows) removed
}
\description{
Filter by removal of one or more instances of values found in one or more variables.
}
