% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepairs.R
\name{verify_filepairs}
\alias{verify_filepairs}
\title{Verify that a list of file pairs is suitable for processing.}
\usage{
verify_filepairs(x, ...)
}
\arguments{
\item{x}{a list with file pairings as character vectors}

\item{...}{arguments for \code{\link[auntie]{verify_filepairs}}}
}
\value{
see \code{\link[auntie]{verify_filepairs}}
}
\description{
Verify that a list of file pairs is suitable for processing.
}
