% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepairs.R
\name{verify_filepairs}
\alias{verify_filepairs}
\title{Verify that a list of file pairs is suitable for processing.}
\usage{
verify_filepairs(
  x,
  min_size = NULL,
  elements = c("forward", "reverse"),
  require_reverse = FALSE
)
}
\arguments{
\item{x}{a list with file pairings as character vectors}

\item{min_size}{numeric, files with fewer bytes than this number are considered empty. Set to NA or NULL to ignore file size.}

\item{elements}{character, the names of the file pair elements to test}

\item{require_reverse}{logical, if TRUE then insist that reverse files must be present (ala Illumina). If FALSE (the default) then allow reverse files to be absent (ala PacBio)}
}
\value{
the input list with possibly all elements removed if all files fail the size test
}
\description{
Verify that a list of file pairs is suitable for processing.
}
