% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepairs.R, R/misc.R
\name{verify_filepairs}
\alias{verify_filepairs}
\title{Verify that a list of file pairs is suitable for processing.}
\usage{
verify_filepairs(
  x,
  elements = c("forward", "reverse"),
  require_reverse = FALSE
)

verify_filepairs(
  x,
  elements = c("forward", "reverse"),
  require_reverse = FALSE
)
}
\arguments{
\item{x}{a list with file pairings as character vectors}

\item{elements}{character, the names of the file pair elements to test}

\item{require_reverse}{logical, if TRUE then insist that reverse files must be present (ala Illumina).
If FALSE (the default) then allow reverse files to be absent (ala PacBio)}
}
\value{
the input list

the input list
}
\description{
This will throw an error if there is a mismatch between the number of
elements in the list elements. Otherwise it simply passes the data through.

This will throw an error if there is a mismatch between the number of
elements in the list elements. Otherwise it simply passes the data through.
}
