% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_QPP}
\alias{write_QPP}
\title{Write quality_profile_pairs object}
\usage{
write_QPP(x, filename = "qpp")
}
\arguments{
\item{x}{output from \code{\link{quality_profile_pairs}}}

\item{filename}{the name of the file to write}
}
\value{
the input errors object
}
\description{
Write quality_profile_pairs object
}
